requirements.txt
fastapi
uvicorn[standard]
httpx
pdf2image
opencv-python
Pillow
python-multipart
python-dotenv
google-genai
